@@ -24,11 +24,12 @@ dependencies:
   flutter:
     sdk: flutter
   url_launcher: ^6.0.0-nullsafety.1
-  font_awesome_flutter: ^9.2.0
+  font_awesome_flutter: ^10.7.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  url_strategy: ^0.2.0
 
 dev_dependencies:
   flutter_test:
@@ -48,6 +49,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - images/profilePhoto.webp
+    - images/appstore.jpg
     - images/applogo.webp
     - images/navibar.gif
 
