@@ -1,9 +1,9 @@
 import 'package:url_launcher/url_launcher.dart';
 
-void launchURL(String socialMediaURL) async {
-  String url = socialMediaURL;
-  if (await canLaunch(url)) {
-    await launch(url);
+void launchURL(String? socialMediaURL) async {
+  String? url = socialMediaURL;
+  if (await canLaunchUrl(Uri.parse(url!))) {
+    await launchUrl(Uri.parse(url));
   } else {
     throw 'Could not launch $url';
   }
