@@ -7,7 +7,38 @@ class ProjectsList {
 
   List<Widget> projectsList = [
     ProjectCard(
-      index: 0,
+      title: 'Bedtime Story AI', 
+      description: 'Bedtime Story AI is an app that allows you to type in a title and procedes to generate a short story for you.', 
+      preview: CircleAvatar(
+        backgroundImage: AssetImage('images/appstore.jpg'),
+        backgroundColor: Colors.black26,
+        radius: 45,
+      ), 
+      linkButtonArea: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            CustomScaleTransition(
+              index: 2,
+              child: LinkButton(
+                buttonText: 'Android',
+                url: '',
+              ),
+            ),
+            SizedBox(width: 10),
+            CustomScaleTransition(
+              index: 3,
+              child: LinkButton(
+                buttonText: 'iOS',
+                url: '',
+              ),
+            ),
+          ],
+        ),
+      index: 0, 
+      keywordStrings: ['flutter', 'supabase', 'chatgpt api'],
+      ),
+    ProjectCard(
+      index: 1,
         title: 'BabyGrowth app',
         description: 'BabyGrowth allows you to input the height and weight data of your children and create their growth charts. An easy way to make colourful charts and share it friends & family.',
         preview: CircleAvatar(
@@ -38,7 +69,7 @@ class ProjectsList {
       keywordStrings: ['flutter', 'firebase', 'user preferences'],
     ),
     ProjectCard(
-      index: 1,
+      index: 2,
       title: 'NaviBar package',
       description: 'A customizable bottom navigation bar for Flutter using simple animations and theme colors that comes with a couple of presets.',
       preview: Image.asset("images/navibar.gif", width: 360, height: 80),
@@ -56,6 +87,6 @@ class ProjectsList {
         description: 'Newly designed personal website using the power of Flutter web',
         preview: FlutterLogo(size: 125,),
         linkButtonArea: Text(''),
-        index: 2, keywordStrings: ['responsive', 'custom animations', 'flutter web', ])
+        index: 3, keywordStrings: ['responsive', 'custom animations', 'flutter web', ])
   ];
 }
