@@ -3,6 +3,7 @@ import 'package:personal_blog_v2/screens/about.dart';
 import 'package:personal_blog_v2/screens/projects.dart';
 import 'package:personal_blog_v2/screens/standard_content.dart';
 import '../components/profile.dart';
+import '../constants.dart';
 
 class MobileContent extends StatelessWidget {
   const MobileContent({Key? key}) : super(key: key);
@@ -15,8 +16,8 @@ class MobileContent extends StatelessWidget {
           begin: Alignment.topCenter,
           end: Alignment.bottomCenter,
           colors: [
-            Colors.indigoAccent,
-            Colors.redAccent,
+            Constants.kGradientOne,
+            Constants.kGradientTwo,
           ],
           stops: [
             0.1,
