@@ -3,7 +3,7 @@ import 'package:personal_blog_v2/animations/staggered_slide_transition.dart';
 import 'package:personal_blog_v2/components/about_content.dart';
 
 class About extends StatelessWidget {
-  static const String id = 'about';
+  static const String id = '/about';
   const About({Key? key}) : super(key: key);
 
   @override
