@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:personal_blog_v2/methods/responsive.dart';
 import 'package:personal_blog_v2/responsive_content/mobile_content.dart';
 import '../responsive_content/desktop_content.dart';
+import '../responsive_content/tablet_content.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -16,7 +17,7 @@ class _HomePageState extends State<HomePage> {
   Widget build(BuildContext context) {
     return Responsive(
         desktop: DesktopContent(),
-        tablet: MobileContent(),
+        tablet: TabletContent(),
         mobile: MobileContent(),
     );
   }
